size_t

+++ 101-binary_tree_levelorder.c
@@ -9,9 +9,9 @@
  * @func: function
  */
 
-void aux(const binary_tree_t *tree, int i, void (*func)(int))
+void aux(const binary_tree_t *tree, size_t i, void (*func)(int))
 {
-	if((size_t) binary_tree_depth(tree) == i)
+	if(binary_tree_depth(tree) == i)
 		func(tree->n);
 	aux(tree->left, i, func);
 	aux(tree->right, i, func);
